@@ -4,7 +4,7 @@ version: 0.2.0+1
 homepage: https://github.com/gomgom/flutter_language_pickers
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
