@@ -29,7 +29,7 @@ class LanguagePickerDialog extends StatefulWidget {
   /// provided (but see [contentPadding]). If it _is_ null, then an extra 20
   /// pixels of bottom padding is added to separate the [title] from the
   /// [actions].
-  final EdgeInsetsGeometry titlePadding;
+  final EdgeInsetsGeometry? titlePadding;
 
   /// Padding around the content.
 
@@ -50,7 +50,7 @@ class LanguagePickerDialog extends StatefulWidget {
 
   ///Callback that is called with selected item of type Language which returns a
   ///Widget to build list view item inside dialog
-  final ItemBuilder itemBuilder;
+  final ItemBuilder? itemBuilder;
 
   /// The (optional) horizontal separator used between title, content and
   /// actions.
@@ -67,24 +67,24 @@ class LanguagePickerDialog extends StatefulWidget {
   final bool isSearchable;
 
   /// The optional [decoration] of search [TextField]
-  final InputDecoration searchInputDecoration;
+  final InputDecoration? searchInputDecoration;
 
   ///The optional [cursorColor] of search [TextField]
   final Color searchCursorColor;
 
   ///The search empty view is displayed if nothing returns from search result
-  final Widget searchEmptyView;
+  final Widget? searchEmptyView;
 
   /// List of languages available in this picker.
-  final List<Map<String, String>> languagesList;
+  final List<Map<String, String>>? languagesList;
 
   LanguagePickerDialog({
-    Key key,
-    this.onValuePicked,
-    this.title,
+    Key? key,
+    required this.onValuePicked,
+    required this.title,
     this.titlePadding,
     this.contentPadding = const EdgeInsets.fromLTRB(0.0, 12.0, 0.0, 16.0),
-    this.semanticLabel,
+    required this.semanticLabel,
     this.itemBuilder,
     this.isDividerEnabled = false,
     this.divider = const Divider(
@@ -92,7 +92,7 @@ class LanguagePickerDialog extends StatefulWidget {
     ),
     this.isSearchable = false,
     this.searchInputDecoration,
-    this.searchCursorColor,
+    required this.searchCursorColor,
     this.searchEmptyView,
     this.languagesList,
   }) : super(key: key);
@@ -104,8 +104,8 @@ class LanguagePickerDialog extends StatefulWidget {
 }
 
 class SingleChoiceDialogState extends State<LanguagePickerDialog> {
-  List<Language> _allLanguages;
-  List<Language> _filteredLanguages;
+  late List<Language> _allLanguages;
+  late List<Language> _filteredLanguages;
 
   @override
   void initState() {
@@ -134,8 +134,8 @@ class SingleChoiceDialogState extends State<LanguagePickerDialog> {
             children: _filteredLanguages
                 .map((item) => SimpleDialogOption(
                       child: widget.itemBuilder != null
-                          ? widget.itemBuilder(item)
-                          : Text(item.name),
+                          ? widget.itemBuilder!(item)
+                          : Text(item.name!),
                       onPressed: () {
                         widget.onValuePicked(item);
                         Navigator.pop(context);
@@ -163,7 +163,7 @@ class SingleChoiceDialogState extends State<LanguagePickerDialog> {
   _buildTitle() {
     return widget.titlePadding != null
         ? Padding(
-            padding: widget.titlePadding,
+            padding: widget.titlePadding!,
             child: widget.title,
           )
         : widget.title;
@@ -178,8 +178,12 @@ class SingleChoiceDialogState extends State<LanguagePickerDialog> {
         setState(() {
           _filteredLanguages = _allLanguages
               .where((Language language) =>
-                  language.name.toLowerCase().startsWith(value.toLowerCase()) ||
-                  language.isoCode.toLowerCase().startsWith(value.toLowerCase()))
+                  language.name!
+                      .toLowerCase()
+                      .startsWith(value.toLowerCase()) ||
+                  language.isoCode!
+                      .toLowerCase()
+                      .startsWith(value.toLowerCase()))
               .toList();
         });
       },
