@@ -5,31 +5,34 @@ import 'package:flutter/material.dart';
 ///Provides a customizable [DropdownButton] for all languages
 class LanguagePickerDropdown extends StatefulWidget {
   LanguagePickerDropdown(
-      {this.itemBuilder, this.initialValue, this.onValuePicked, this.languagesList});
+      {this.itemBuilder,
+      this.initialValue,
+      required this.onValuePicked,
+      this.languagesList});
 
   ///This function will be called to build the child of DropdownMenuItem
   ///If it is not provided, default one will be used which displays
   ///flag image, isoCode and phoneCode in a row.
   ///Check _buildDefaultMenuItem method for details.
-  final ItemBuilder itemBuilder;
+  final ItemBuilder? itemBuilder;
 
   ///It should be one of the ISO ALPHA-2 Code that is provided
   ///in languagesList map of languages.dart file.
-  final String initialValue;
+  final String? initialValue;
 
   ///This function will be called whenever a Language item is selected.
   final ValueChanged<Language> onValuePicked;
 
   /// List of languages available in this picker.
-  final List<Map<String, String>> languagesList;
+  final List<Map<String, String>>? languagesList;
 
   @override
   _LanguagePickerDropdownState createState() => _LanguagePickerDropdownState();
 }
 
 class _LanguagePickerDropdownState extends State<LanguagePickerDropdown> {
-  List<Language> _languages;
-  Language _selectedLanguage;
+  late List<Language> _languages;
+  late Language _selectedLanguage;
 
   @override
   void initState() {
@@ -38,8 +41,7 @@ class _LanguagePickerDropdownState extends State<LanguagePickerDropdown> {
     if (widget.initialValue != null) {
       try {
         _selectedLanguage = _languages
-            .where((language) =>
-                language.isoCode == widget.initialValue)
+            .where((language) => language.isoCode == widget.initialValue)
             .toList()[0];
       } catch (error) {
         throw Exception(
@@ -58,7 +60,7 @@ class _LanguagePickerDropdownState extends State<LanguagePickerDropdown> {
         .map((language) => DropdownMenuItem<Language>(
             value: language,
             child: widget.itemBuilder != null
-                ? widget.itemBuilder(language)
+                ? widget.itemBuilder!(language)
                 : _buildDefaultMenuItem(language)))
         .toList();
 
@@ -69,7 +71,7 @@ class _LanguagePickerDropdownState extends State<LanguagePickerDropdown> {
             isDense: true,
             onChanged: (value) {
               setState(() {
-                _selectedLanguage = value;
+                _selectedLanguage = value!;
                 widget.onValuePicked(value);
               });
             },
