@@ -1,8 +1,8 @@
 class Language {
   Language(this.isoCode, this.name);
 
-  final String name;
-  final String isoCode;
+  final String? name;
+  final String? isoCode;
 
   Language.fromMap(Map<String, String> map)
       : name = map['name'],
